Add files via upload
Router Scan can find and identify various devices from a large number of known routers / routers and, most importantly, extract useful information from them, in particular wireless network characteristics: access point protection method (encryption), access point name (SSID) and key access points (passphrase). It also gets information about the WAN connection (useful when scanning a local network) and displays the brand and model of the router. Information is obtained in two possible ways: the program will try to guess a login / password pair for the router from the list of standard passwords, as a result of which it will gain access. Or non-destructive vulnerabilities (or bugs) for a specific router model will be used, allowing to obtain the necessary information and / or bypass the authorization process.